Refactor app instantiation
