+++ api/Makefile
@@ -1,0 +1,2 @@
+run-debug-server:
+	FLASK_ENV=development python3 -m flask run
