+++ api/app.py
@@ -1,0 +1,14 @@
+from flask import Flask
+
+from routes.cmc import cmc_api
+
+
+def create_app():
+    app = Flask(__name__)
+
+    app.register_blueprint(cmc_api)
+
+    return app
+
+
+new_app = create_app()
