+++ docker-compose.yml
@@ -2,7 +2,7 @@ version: "3.7"
 services:
   api:
     build: ./api
-    command: "python3 -m gunicorn -b 0.0.0.0 --workers=2 --chdir .. api:new_app"
+    command: "python3 -m gunicorn -b 0.0.0.0 --workers=2 app:new_app"
     expose:
       - 8000
     environment:
